Move source components out of corpus/ and reuse for lexicon
